WebHost: move room_status out of __init__.py
The old location is unexpected and easy to miss.
